# CUDA: add a fused top-K MoE kernel