This kernel does the following:
1. softmax over the logits per token [n_experts, n_tokens]
2. argmax reduce over the top-k (n_experts_used) logits
3. write weights + ids to global memory

It is intended as fusion of softmax->top-k->get_rows pipeline for MoE models

Should be more useful in TG than PP

| Model                   | Test   |   t/s master |   t/s patch |   Speedup |
|:------------------------|:-------|-------------:|------------:|----------:|
| qwen3moe 30B.A3B Q4_K_M | pp512  |      6502.24 |     6561.50 |      1.01 |
| qwen3moe 30B.A3B Q4_K_M | tg128  |       194.63 |      207.75 |      1.07 |