@am17an 2025-09-21
Some more performance numbers 

| Model                   |   Microbatch size | Test   |   t/s master |   t/s patch |   Speedup |
|:------------------------|------------------:|:-------|-------------:|------------:|----------:|
| qwen3moe 30B.A3B Q4_K_M |                 1 | pp4096 |       194.25 |      205.88 |      1.06 |
| qwen3moe 30B.A3B Q4_K_M |                 2 | pp4096 |       206.51 |      213.80 |      1.04 |
| qwen3moe 30B.A3B Q4_K_M |                 4 | pp4096 |       355.94 |      366.72 |      1.03 |
| qwen3moe 30B.A3B Q4_K_M |                 8 | pp4096 |       585.58 |      600.32 |      1.03 |
| qwen3moe 30B.A3B Q4_K_M |                16 | pp4096 |       922.33 |      940.61 |      1.02 |
| qwen3moe 30B.A3B Q4_K_M |                32 | pp4096 |      1533.54 |     1564.10 |      1.02 |
| qwen3moe 30B.A3B Q4_K_M |               256 | pp4096 |      3722.88 |     3751.81 |      1.01 |
| qwen3moe 30B.A3B Q4_K_M |               512 | pp4096 |      4559.34 |     4588.63 |      1.01 |