@am17an 2025-09-24
It does not crash when doing `--no-warmup`, somehow setting n_expert_used = n_expert during warmup triggers an illegal memory access, however doing this in `test-backend-ops` does not, and `compute-sanitizer` is also clean. I don't see anything that changes in warmup except for setting n_expert_used = n_expert, any help would be appreciated!