@am17an 2025-09-24
> Does it still happen if you keep the build forward expand and remove the new fusing logic?

No doesn't happen if I remove the fusing logic and keep the build forward expand