@ggerganov 2025-09-24
> Maybe we can just remove no-ops from the graphs node lists. It would require changes in `ggml-alloc`, however.

If this has no side-effects on the scheduler/allocator logic, would be the best option. I don't think the backends would ever need to see the empty nodes of the graph - they should always need only the nodes that do actual reads and writes.