@am17an 2025-09-24
> If this has no side-effects on the scheduler/allocator logic, would be the best option. I don't think the backends would ever need to see the empty nodes of the graph - they should always need only the nodes that do actual reads and writes.

This would complicate fusion at least. Right now you can look at llama-graph and call build forward to get the exact sequence you see in llama-graph. I don't know how it would look like once these ops go